requirements.txt
tk
ttkbootstrap
pyautogui
pytesseract
pyperclip
Pillow
screeninfo
keyboard
pystray